package정리, Long->Integer로 변경, Post 기본적인거 초안 생성

See merge request tkdtkd975/final_project_leesanghun_team2!20

+++ src/main/java/com/final_project_leesanghun_team2/repository/UserRepository.java
@@ -1,10 +1,10 @@
 package com.final_project_leesanghun_team2.repository;
 
-import com.final_project_leesanghun_team2.domain.User;
+import com.final_project_leesanghun_team2.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUserName(String userName);
 }
